restructure tests to support easier version thinking
