+++ schemas/v2/index.js
@@ -1,9 +1,5 @@
-const ritual = require('./ritual')
-const root = require('./root')
 const shard = require('./shard')
 
 module.exports = {
-  ritual,
-  root,
   shard
 }
