--- schemas/v2/ritual.js
@@ -1,30 +1,0 @@
-const definitions = require('ssb-schema-definitions')
-const SCHEMA_VERSION = require('./version')
-
-module.exports = {
-  $schema: 'http://json-schema.org/schema#',
-  type: 'object',
-  required: ['type', 'version', 'root', 'quorum', 'shards', 'tool'],
-  properties: {
-    type: {
-      type: 'string',
-      pattern: '^dark-crystal/ritual$'
-    },
-    version: {
-      type: 'string',
-      pattern: `^${SCHEMA_VERSION}$`
-    },
-    root: { $ref: '#/definitions/messageId' },
-    quorum: {
-      type: 'integer',
-      minimum: 2
-    },
-    shards: {
-      type: 'integer',
-      minimum: 2
-    },
-    tool: { type: 'string' },
-    recps: { $ref: '#/definitions/recps' }
-  },
-  definitions: definitions
-}
