+++ schemas/v2/version.js
@@ -1,2 +1,2 @@
-const SCHEMA_VERSION = '2.0.0'
-module.exports = SCHEMA_VERSION
+// SCHEMA_VERSION
+module.exports = '2.0.0'
